+++ GPU (CUDA)/include/sphere.h
@@ -6,10 +6,10 @@
 
 struct Sphere : Shape
 {
-    double r;
+    float r;
     Vec c;
-    __device__ Sphere(double _r, Vec _c);
-    __device__ bool Intersect(Ray &ray, double &hit) override;
+    __device__ Sphere(float _r, Vec _c);
+    __device__ bool Intersect(Ray &ray, float &hit) override;
     __device__ Vec Normal(Vec &v) override;
     __device__ void Translate(Vec &x) override;
 };
