+++ GPU (CUDA)/src/scene.cu
@@ -5,9 +5,9 @@
 __device__ HitRecord Scene::ClosestIntersection(Ray r, Solid **objects, int nObj)
 {
     // Assume no hit initally
-    double recordT = PTUtility::INF;
+    float recordT = PTUtility::INF;
     int idMin = -1;
-    double t;
+    float t;
 
     // iterate through and check for a hit
     for (long unsigned i = 0; i < nObj; i++)
@@ -29,11 +29,11 @@ __device__ HitRecord Scene::ClosestIntersection(Ray r, Solid **objects, int nObj
 
 __device__ Ray Scene::GenerateDiffuseRay(Vec &hitPt, Vec &normal, curandState &state)
 {
-    double phi = 2 * PTUtility::PI * PTUtility::Random(state);
-    double r2 = PTUtility::Random(state);
+    float phi = 2 * PTUtility::PI * PTUtility::Random(state);
+    float r2 = PTUtility::Random(state);
 
-    double sinTheta = sqrt(r2);
-    double cosTheta = sqrt(1 - r2);
+    float sinTheta = sqrt(r2);
+    float cosTheta = sqrt(1 - r2);
 
     // building basis for R3 of perpendicular vectors around the normal facing outwards
     Vec w = normal.Norm();
@@ -51,17 +51,17 @@ __device__ Ray Scene::GenerateSpecularRay(Vec &hitPt, Vec &rayNormal, Vec &dir)
     return Ray(hitPt, reflectedDir);
 }
 
-__device__ Vec Scene::Jitter(Vec x, double r, curandState &state)
+__device__ Vec Scene::Jitter(Vec x, float r, curandState &state)
 {
     Vec w = x.Norm();
     Vec u = Vec(w.z, w.z, -w.x - 2 * w.y).Norm();
     Vec v = Vec::Cross(u, w).Norm();
 
-    double rad = r * sqrt(PTUtility::Random(state));
-    double theta = PTUtility::Random(state) * 2 * PTUtility::PI;
+    float rad = r * sqrt(PTUtility::Random(state));
+    float theta = PTUtility::Random(state) * 2 * PTUtility::PI;
 
-    double xPrime = rad * cos(theta);
-    double yPrime = rad * sin(theta);
+    float xPrime = rad * cos(theta);
+    float yPrime = rad * sin(theta);
 
     return (u * xPrime + v * yPrime + w).Norm();
 }
@@ -118,19 +118,19 @@ __device__ Vec Scene::RayColour(Ray r, Solid **objects, curandState &state, int 
             bool isInto = Vec::Dot(normal, rayNormal) > 0;
 
             // refractive indices
-            double n1 = 1;
-            double n2 = 1.5;
-            double netN = isInto ? n1 / n2 : n2 / n1;
+            float n1 = 1;
+            float n2 = 1.5;
+            float netN = isInto ? n1 / n2 : n2 / n1;
 
             // cosines of the angles
-            double cosTheta = Vec::Dot(r.d, rayNormal);
-            double cosTheta2Sqr = 1 - netN * netN * (1 - cosTheta * cosTheta);
+            float cosTheta = Vec::Dot(r.d, rayNormal);
+            float cosTheta2Sqr = 1 - netN * netN * (1 - cosTheta * cosTheta);
 
             // total internal reflection
             if (cosTheta2Sqr < 0)
             {
                 r = reflectedRay;
-                if(hit->s == Surface::REFRGLOSS)
+                if (hit->s == Surface::REFRGLOSS)
                     r.d = Scene::Jitter(r.d, 0.125, state);
             }
             else
@@ -140,27 +140,26 @@ __device__ Vec Scene::RayColour(Ray r, Solid **objects, curandState &state, int 
                 // Vec refractedDir = (r.d * netN - rayNormal * (cosTheta * cosTheta + sqrt(cosTheta2Sqr)))).Norm();
 
                 // approximating reflection and refraction weights
-                double a = n2 - n1;
-                double b = n1 + n2;
-                double R0 = (a * a) / (b * b);
+                float a = n2 - n1;
+                float b = n1 + n2;
+                float R0 = (a * a) / (b * b);
 
-                double cosTheta2 = Vec::Dot(refractedDir, normal);
-                double c = 1 - (isInto ? -cosTheta : cosTheta2);
-                double refl = R0 + (1 - R0) * c * c * c * c * c;
-                double refr = 1 - refl;
+                float cosTheta2 = Vec::Dot(refractedDir, normal);
+                float c = 1 - (isInto ? -cosTheta : cosTheta2);
+                float refl = R0 + (1 - R0) * c * c * c * c * c;
+                float refr = 1 - refl;
 
-                double P = 0.25 + 0.5 * refl;
+                float P = 0.25 + 0.5 * refl;
 
-                double reflectionWeight = refl / P;
-                double refractionWeight = refr / (1 - P);
+                float reflectionWeight = refl / P;
+                float refractionWeight = refr / (1 - P);
 
                 if (PTUtility::Random(state) < P)
                 {
                     awayMask = awayMask * reflectionWeight;
                     r = reflectedRay;
-                    if(hit->s == Surface::REFRGLOSS)
+                    if (hit->s == Surface::REFRGLOSS)
                         r.d = Scene::Jitter(r.d, 0.125, state);
-                        
                 }
                 else
                 {
